@@ -9,13 +9,13 @@
     <version>1.0.0-SNAPSHOT</version>
 
     <properties>
-        <maven.compiler.target>1.8</maven.compiler.target>
-        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <maven.compiler.source>17</maven.compiler.source>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 
         <springfox.version>3.0.0</springfox.version>
         <archunit.version>1.0.1</archunit.version>
-        <spring.boot.version>2.3.8.RELEASE</spring.boot.version>
+        <spring.boot.version>3.0.0</spring.boot.version>
     </properties>
 
     <dependencyManagement>
