@@ -10,7 +10,7 @@
     <name>craftsman.all</name>
 
     <properties>
-        <maven.compiler.source>1.8</maven.compiler.source>
+        <maven.compiler.source>17</maven.compiler.source>
         <maven.compiler.target>${maven.compiler.source}</maven.compiler.target>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
@@ -18,8 +18,7 @@
 
         <cola.components.version>4.4.0-SNAPSHOT</cola.components.version>
 
-        <spring-boot.version>2.7.5</spring-boot.version>
-        <mybatis-starter.version>2.2.2</mybatis-starter.version>
+        <spring-boot.version>3.0.0</spring-boot.version>
     </properties>
 
     <modules>
@@ -91,7 +90,7 @@
             <dependency>
                 <groupId>org.mybatis.spring.boot</groupId>
                 <artifactId>mybatis-spring-boot-starter</artifactId>
-                <version>${mybatis-starter.version}</version>
+                <version>${spring-boot.version}</version>
             </dependency>
 
             <!--Validation API-->
@@ -115,6 +114,11 @@
                 <version>1.2.83</version>
             </dependency>
             <!-- Misc End -->
+            <dependency>
+                <groupId>org.glassfish</groupId>
+                <artifactId>jakarta.el</artifactId>
+                <version>3.0.4</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
 
