@@ -5,7 +5,7 @@ package ${package}.dto.data;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotEmpty;
 
 @Data
 public class CustomerDTO{
