@@ -12,7 +12,7 @@ import com.huawei.charging.domain.gateway.AccountGateway;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.ArrayList;
 import java.util.List;
 
