@@ -4,7 +4,7 @@ import com.huawei.charging.domain.charge.Session;
 import com.huawei.charging.domain.gateway.AccountGateway;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 @Component
 public class AccountDomainService {
