@@ -9,7 +9,7 @@ import com.alibaba.craftsman.gatewayimpl.database.dataobject.MetricDO;
 import com.alibaba.fastjson.JSON;
 import org.springframework.stereotype.Component;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 import java.util.ArrayList;
 import java.util.List;
 
