@@ -3,7 +3,7 @@ package com.alibaba.craftsman.dto;
 import com.alibaba.craftsman.dto.clientobject.UserProfileCO;
 import lombok.Data;
 
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 @Data
 public class UserProfileUpdateCmd extends CommonCommand {
