@@ -2,7 +2,7 @@ package com.alibaba.craftsman.dto.clientobject;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotEmpty;
 
 /**
  * PaperMetricCO
